requirements.txt

Copy
flask
flask-sqlalchemy
flask-jwt-extended
python-dotenv
werkzeug